# omakase/src/main/webapp/swagger/v1/omakase.json
{
  "basePath": "/omakase/api",
  "swagger": "2.0",
  "info": {
    "title": "Omakase",
    "version": "1.0"
  },
...
  "paths": {
...
    "/assets/{assetId}/variants/{variantId}/repositories/{repositoryId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "getRepository",
        "summary": "Retrieves a variant's repository",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "Variant Id",
            "name": "variantId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "Repository Id",
            "name": "repositoryId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetVariantRepository"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "assets"
        ],
        "operationId": "deleteVariantFromRepository",
        "summary": "Deletes a variant from a repository",
        "parameters": [
          {
            "in": "path",
            "description": "Asset Id",
            "name": "assetId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "Variant Id",
            "name": "variantId",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "description": "Repository Id",
            "name": "repositoryId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseStatus"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "jobs"
        ],
        "operationId": "getJobs",
        "summary": "Retrieve a listing of jobs",
        "parameters": [
          {
            "in": "query",
            "description": "The page of results to return",
            "name": "page",
            "required": false,
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "description": "The number of results per page",
            "name": "per_page",
            "required": false,
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "query",
            "description": "Determines the attribute used to sort the results",
            "name": "sort",
            "required": false,
            "type": "string",
            "default": "created"
          },
          {
            "in": "query",
            "description": "Determines the sort order of the results",
            "name": "order",
            "required": false,
            "type": "string",
            "default": "DESC"
          },
          {
            "in": "query",
            "description": "Limits the results to an aggregate, pagination is ignored.",
            "name": "only_count",
            "required": false,
            "type": "boolean",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/JobsEnvelope"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "jobs"
        ],
        "operationId": "createJob",
        "summary": "Create a new job",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostJob"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ResponseStatus"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/repositories/{repositoryId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "repositories"
        ],
        "operationId": "getRepository",
        "summary": "Retrieve a repository",
        "parameters": [
          {
            "in": "path",
            "description": "Repository Id",
            "name": "repositoryId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetRepository"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "repositories"
        ],
        "operationId": "deleteRepository",
        "summary": "Delete a repository",
        "parameters": [
          {
            "in": "path",
            "description": "Repository Id",
            "name": "repositoryId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseStatus"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/status/{statusId}": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "status"
        ],
        "operationId": "getStatus",
        "summary": "Retrieves a status",
        "parameters": [
          {
            "in": "path",
            "description": "Status Id",
            "name": "statusId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Status"
            }
          },
          "401": {
            "description": "Not Authorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "ResponseStatus": {
      "required": [
        "status",
        "links"
      ],
      "properties": {
        "status": {
          "type": "string",
          "enum": [
            "OK",
            "ERROR"
          ]
        },
        "message": {
          "type": "string"
        },
        "exception": {
          "type": "string"
        },
        "links": {
          "$ref": "#/definitions/ResponseStatusLinks"
        }
      }
    },
    "ResponseStatusLinks": {
      "required": [
        "status"
      ],
      "properties": {
        "status": {
          "$ref": "#/definitions/Href"
        }
      }
    },
    "Envelope": {
      "required": [
        "total_count",
        "data"
      ],
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "per_page": {
          "type": "integer",
          "format": "int32"
        },
        "total_pages": {
          "type": "integer",
          "format": "int64"
        },
        "total_count": {
          "type": "integer",
          "format": "int64"
        },
        "links": {
          "$ref": "#/definitions/PaginationLinks"
        }
      }
    },
    "PaginationLinks": {
      "properties": {
        "first": {
          "$ref": "#/definitions/Href"
        },
        "last": {
          "$ref": "#/definitions/Href"
        },
        "prev": {
          "$ref": "#/definitions/Href"
        },
        "next": {
          "$ref": "#/definitions/Href"
        }
      }
    },
    "Href": {
      "required": [
        "href"
      ],
      "properties": {
        "href": {
          "type": "string"
        }
      }
    },
...
    "GetVariantRepository": {
      "required": [
        "id",
        "name",
        "type",
        "created",
        "created_by",
        "last_modified",
        "last_modified_by",
        "links"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "created_by": {
          "type": "string"
        },
        "last_modified": {
          "type": "string",
          "format": "date-time"
        },
        "last_modified_by": {
          "type": "string"
        },
        "links": {
          "$ref": "#/definitions/VariantRepositoryLinks"
        }
      }
    },
    "VariantRepositoryLinks": {
      "required": [
        "self",
        "configuration",
        "template"
      ],
      "properties": {
        "self": {
          "$ref": "#/definitions/Href"
        },
        "configuration": {
          "$ref": "#/definitions/Href"
        },
        "template": {
          "$ref": "#/definitions/Href"
        }
      }
    },
...
    "JobsEnvelope": {
      "allOf": [
        {
          "$ref": "#/definitions/Envelope"
        },
        {
          "type": "object",
          "properties": {
            "data": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/GetJob"
              }
            }
          }
        }
      ]
    },
    "GetJob": {
      "required": [
        "id",
        "type",
        "status",
        "priority",
        "created",
        "created_by",
        "last_modified",
        "last_modified_by",
        "links"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "external_ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "type": {
          "type": "string",
          "enum": [
            "INGEST",
            "EXPORT",
            "REPLICATE",
            "TRANSFORMATION"
          ]
        },
        "status": {
          "$ref": "#/definitions/CurrentJobStatus"
        },
        "priority": {
          "type": "integer",
          "format": "int64",
          "default": 4
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "created_by": {
          "type": "string"
        },
        "last_modified": {
          "type": "string",
          "format": "date-time"
        },
        "last_modified_by": {
          "type": "string"
        },
        "links": {
          "$ref": "#/definitions/JobLinks"
        }
      }
    },
    "CurrentJobStatus": {
      "required": [
        "current",
        "timestamp"
      ],
      "properties": {
        "current": {
          "type": "string",
          "enum": [
            "UNSUBMITTED",
            "QUEUED",
            "EXECUTING",
            "COMPLETED",
            "FAILED",
            "CANCELED"
          ]
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "JobLinks": {
      "required": [
        "self",
        "status",
        "configuration"
      ],
      "properties": {
        "self": {
          "$ref": "#/definitions/Href"
        },
        "status": {
          "$ref": "#/definitions/Href"
        },
        "configuration": {
          "$ref": "#/definitions/Href"
        }
      }
    },
    "PostJob": {
      "required": [
        "type",
        "configuration"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "external_ids": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "type": {
          "type": "string",
          "enum": [
            "INGEST",
            "EXPORT",
            "REPLICATE",
            "TRANSFORMATION"
          ]
        },
        "status": {
          "type": "string",
          "enum": [
            "UNSUBMITTED",
            "QUEUED"
          ]
        },
        "priority": {
          "type": "integer",
          "format": "int64",
          "default": 4
        },
        "configuration": {
          "$ref": "#/definitions/JobConfiguration"
        }
      }
    },
    "JobConfiguration": {},
...
    "GetRepository": {
      "required": [
        "id",
        "name",
        "type",
        "created",
        "created_by",
        "last_modified",
        "last_modified_by"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "created_by": {
          "type": "string"
        },
        "last_modified": {
          "type": "string",
          "format": "date-time"
        },
        "last_modified_by": {
          "type": "string"
        },
        "links": {
          "$ref": "#/definitions/RepositoryLinks"
        }
      }
    },
    "RepositoryLinks": {
      "required": [
        "self",
        "configuration",
        "template"
      ],
      "properties": {
        "self": {
          "$ref": "#/definitions/Href"
        },
        "configuration": {
          "$ref": "#/definitions/Href"
        },
        "template": {
          "$ref": "#/definitions/Href"
        }
      }
    },
...
    "Status": {
      "required": [
        "current",
        "timestamp",
        "links"
      ],
      "properties": {
        "current": {
          "type": "string",
          "enum": [
            "QUEUED",
            "EXECUTING",
            "COMPLETED",
            "FAILED",
            "CANCELED"
          ]
        },
        "timestamp": {
          "type": "string",
          "format": "date-time"
        },
        "links": {
          "$ref": "#/definitions/StatusLinks"
        }
      }
    },
    "StatusLinks": {
      "required": [
        "job"
      ],
      "properties": {
        "job": {
          "$ref": "#/definitions/Href"
        }
      }
    },
...
  }
}